Refactor module resposibilities
The MPEG.TS.Demuxer is responsible of buffering and demuxing payloads.
It does so with the help of a new RingBuffer that discards the oldest
packets when the buffer is full. This way, we can bound its usage while
still accepting packets from multiple streams. Users can later subscribe
to a stream track.